server/pyproject.toml
[tool.poetry]
name = "grpc-server"
version = "0.1.0"
description = "A gRPC server project"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
asyncpg = "^0.27.0"
aerich = "^0.7.2"
command = "^0.1.0"
protobuf = "5.28.2" 
pydantic = "^2.9.2"
grpcio-tools = "^1.66.2"
grpcio = "^1.66.2"
structlog = "^24.4.0"
asyncodbc = "^0.1.1"
cryptography = "^43.0.1"

cx-Oracle = "^8.0.0"
tortoise-orm = {extras = ["asyncodbc"], version = "^0.21.6"}
oracledb = "^2.4.1"
python-dateutil = "^2.9.0.post0"




[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"